third_party/blink/web_tests/platform/generic/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/RollbackEvents.https-expected.txt

This is a testharness.js-based test.
FAIL [audio] Track with stream: removal due to disassociation in rollback and then add it back again assert_equals: expected 1 but got 0
FAIL [audio] Track without stream: removal due to disassociation in rollback and then add it back assert_equals: expected 1 but got 0
FAIL [audio] Track with stream: removal due to direction changing and then add back using rollback promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'direction')"
FAIL [audio] Track without stream: removal due to direction changing and then add back using rollback promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'direction')"
FAIL [video] Track with stream: removal due to disassociation in rollback and then add it back again assert_equals: expected 1 but got 0
FAIL [video] Track without stream: removal due to disassociation in rollback and then add it back assert_equals: expected 1 but got 0
FAIL [video] Track with stream: removal due to direction changing and then add back using rollback promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'direction')"
FAIL [video] Track without stream: removal due to direction changing and then add back using rollback promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'direction')"
Harness: the test ran to completion.
